Implement test cases for different problems, modified program and test structure accordingly. Added documentation

+++ Cpp_hands_on_programming/lib/abstract/Ipc_method.cc
@@ -2,7 +2,7 @@
 #include <sys/stat.h>
 #include <iostream>
 
-off_t Ipc_method::getFileSize(const std::string &filename) {
+int Ipc_method::getFileSize(const std::string &filename) {
 	// Use stat to find the size of the file
 	struct stat st;
 	if (stat(filename.c_str(), &st) == -1) {
