+++ Cpp_hands_on_programming/lib/abstract/Ipc_method.h
@@ -11,7 +11,7 @@ class Ipc_method{
 public:
     virtual void send() = 0;
     virtual void receive() = 0;
-    off_t getFileSize(const std::string &filename);
+    int getFileSize(const std::string &filename);
     int readFile(const std::string &filename, int offset, std::vector<char> &buffer, int size);
     int writeFile(const std::string &filename, int total, const std::vector<char> &buffer);
     virtual ~Ipc_method() = default;
